requirements.txt
attrs==20.3.0
fpdf2==2.3.0
iniconfig==1.1.1
packaging==20.9
Pillow==8.1.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.15.0
toml==0.10.2
